Update test.js
The environment variable "test" should be set above the line which includes the index module. If not, the require starts executing the index.js and since the environment variable is not set to test it would start calling the functions as the if check would succeed in index.js. As a result of which the functions would be called twice one via the require index statement and one via test.js. This leads to the failure of createRepo function as one thread would create a repo and the second thread would fail during it's creation with the error - name already exists